+++ src/ENode/Eventing/IEventPublishInfoStore.cs
@@ -9,7 +9,7 @@
         void InsertFirstPublishedVersion(string aggregateRootId);
         /// <summary>Update the published event version of aggregate.
         /// </summary>
-        void UpdatePublishedVersion(string aggregateRootId, long version);
+        void UpdatePublishedVersion(string aggregateRootId, int version);
         /// <summary>Get the current event published version for the specified aggregate.
         /// </summary>
         long GetEventPublishedVersion(string aggregateRootId);
